Tests: Setup (#23)
### Features
- Add testcontainers dependencies;
- Add `image` field to docker compose;
- Add `spring-security-test` as dependency;
- Add `IntegrationTestConfig`, `UnitTestConfig` and `RepositoryUnitTestConfig` for test setup;
- Add `DatabaseContainerInitializer` to fake the database;
- Add password and role validation to `/auth/authenticate`;
- Add application-test.properties;
- Add getter and setter to `registrationIsEnable`;

### Fixes
- Fix DELETE not checking if item exists;
- Fix wrong `/auth/register` status code;
- Fix wrong save status code in all controllers;

### Changes
- Delete HawAPIApplicationTests;
- Replace byte to Byte in some models;
- Update MappersConfig (Add more configurations);

+++ src/main/java/com/lucasjosino/hawapi/services/CharacterService.java
@@ -78,6 +78,8 @@ public class CharacterService {
 
     @Transactional
     public void delete(UUID uuid) {
+        if (!characterRepository.existsById(uuid)) throw new ItemNotFoundException();
+
         characterRepository.deleteById(uuid);
     }
 }
